GPS AF and PF updated
Right now uses onResume method for each fragment to start a refresh. I
need it to do so when the tab/fragment is reopened/activated. Despite my
attempts I wasn't sure how to do so as I didn't do that part of the ui.
The code from onResume(s) can be transfered directly to such an
implementation. Alternatively we can make a button which may be
simpler/safer/faster (my preference)